+++ database/migrations/2023_03_31_181650_create_actor_film_table.php
@@ -12,9 +12,12 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('actor_film', function (Blueprint $table) {
-            $table->integer('actor_id');
-            $table->integer('film_id');
+            $table->unsignedBigInteger('actor_id');
+            $table->unsignedBigInteger('film_id');
             $table->timestamps();
+
+            //$table->foreign('actor_id')->references('id')->on('actors');
+            //$table->foreign('film_id')->references('id')->on('films');
         });
     }
